pyproject.toml
[tool.poetry]
name = "regtraka"
version = "0.1.0"
description = ""
authors = ["AbdulAfeez <abdulafeez.adeyemo@kibo.school>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"
drf-yasg = {extras = ["validation"], version = "^1.21.7"}
pillow = "^10.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
